fixed bugs from making image browser a separate component
explicitly pass whether we are in the saved to the image browser. Maybe not the best solution, but works for now. Also, fixed dependencies of useEffect in saved to prevent constant network calls. Additional fixed some other minor bugs with ImagePreview not showing correct buttons and using index as key causing weird component update behavior on image deletion.